Welcome to the website for BetterThanMinecon 2016 2.0 - the irregular Minecraft modding convention... in Minecraft!
BTM 2016 2.0 is an online Minecraft modding convention made in Minecraft itself. It's also the best place to meet your favourite modders and check out their mods! Voice chat and other facilities will be provided just for that. You'll be able to see mod booths for both the more and less popular mods, as well as check out panels on a variety of subjects! On top of that, you can expect a lot of fun, silliness and surprises!